R25-h: >>1 单位错误族 22 处归正 + ImportEnvironment 孤儿真体接线, 63/63 三组零回归
- 审计不变量: dc 全库 0 处裸指针相减(一切差值=字节域), msvc 元素差>>1 必折半(analysis/r25h)
- 组1: execline 双展开器 7 处(含出口返回值双重复偿) + strbld FormatTypedMemValue 字节寻址
- 组2: b8m 浮点/LIST 3 + b1_remaining 卷枚举/图标/关联链 6 + scriptrun RUN 前缀 1
- 组3: 恒0壳删除, FUN_14007BF78 改名 PECMD_ImportEnvironment 真体接线(i28c '&' 分支)
- divergences D-18(67 动词登记)/D-19(-mode 恒死)/D-20(*map: 桩)/D-21(run_case 时序)
- 每组独立构建+全量 63/63 零回归(dd5d5bdc/e6ecc49d/d05def76)

+++ src/lang/core_scriptrun.c
@@ -310,7 +310,8 @@ int64_t PECMD_RunCommand(void *script, WCHAR *cmdline)
         if (*pt2 == L'"' && pt2[1] == L'.' && (uint16_t)pt2[2] == 0x22 && pt2[3] == L'#') {
             /* ".# 组合 → g_hInstance 模式（TODO(verify): 原语义为版本串检测） */
             pt2 = pt2 + 3;
-            lVar32 = (int64_t)(pt2 - p) >> 1;
+            /* dc:29657 字节差>>1 ≡ 元素数 (R25-h #5) */
+            lVar32 = (int64_t)((intptr_t)pt2 - (intptr_t)p) >> 1;
         }
 
         /* 170-173：sysinit_end 检查 */
